pyproject.toml
[tool.poetry]
name = "python-opsi-common"
version = "4.2.0.1"
description = "The opsi python common library"
homepage = "https://www.opsi.org"
license = "AGPL-3.0"
maintainers = ["uib GmbH <info@uib.de>"]
authors = ["uib GmbH <info@uib.de>"]
packages = [
    { include = "opsicommon" }
]

[[tool.poetry.source]]
name = "uibpypi"
url = "http://pypi.uib.gmbh:8080/simple"

[tool.poetry.dependencies]
python = "^3.7"
colorlog = "^4.1.0"
psutil = "^5.7.0"
pycryptodome = "^3.10.1"
msgpack = "^1.0.2"
requests = "^2.26.0"
lz4 = "^3.1.3"

[tool.poetry.dev-dependencies]
opsi-dev-tools = "^1.0"
pytest = "^5.4.3"
pytest-cov = "^2.9.0"
pylint = "^2.5.2"
flake8 = "^3.8.2"
python-opsi = "^4.2.0.23"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
